requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
redis==6.4.0
apscheduler==3.11.0
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
pytz==2025.2
requests==2.32.5
pandas==2.3.2
python-dotenv==1.0.1
